refactor: borrow incentive pools structure
